+++ frontend/src/metabase/common/components/DatePicker/RelativeDatePicker/DateIntervalPicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./DateIntervalPicker";
