+++ frontend/src/metabase/common/components/DatePicker/RelativeDatePicker/RelativeDatePicker.tsx
@@ -1,33 +1,11 @@
 import { useState } from "react";
-import { t } from "ttag";
-import { Icon } from "metabase/core/components/Icon";
-import {
-  Button,
-  Divider,
-  Flex,
-  Group,
-  Menu,
-  NumberInput,
-  Select,
-  Tabs,
-} from "metabase/ui";
+import { Divider, Group, Tabs } from "metabase/ui";
 import { BackButton } from "../BackButton";
 import type { RelativeDatePickerValue } from "../types";
 import { DEFAULT_VALUE, TABS } from "./constants";
-import {
-  describeInterval,
-  getDirection,
-  getIncludeCurrent,
-  getInterval,
-  getUnitOptions,
-  isIntervalValue,
-  setDirection,
-  setIncludeCurrent,
-  setInterval,
-  setUnit,
-} from "./utils";
-import type { RelativeDateIntervalValue } from "./types";
+import { getDirection, isIntervalValue, setDirection } from "./utils";
 import { CurrentDatePicker } from "./CurrentDatePicker";
+import { DateIntervalPicker } from "./DateIntervalPicker";
 import { TabList } from "./RelativeDatePicker.styled";
 
 interface RelativeDatePickerProps {
@@ -72,7 +50,7 @@ export function RelativeDatePicker({
       {TABS.map(tab => (
         <Tabs.Panel key={tab.direction} value={tab.direction}>
           {isIntervalValue(value) ? (
-            <IntervalPicker
+            <DateIntervalPicker
               value={value}
               isNew={isNew}
               onChange={setValue}
@@ -84,82 +62,5 @@ export function RelativeDatePicker({
         </Tabs.Panel>
       ))}
     </Tabs>
-  );
-}
-
-interface IntervalPickerProps {
-  value: RelativeDateIntervalValue;
-  isNew: boolean;
-  onChange: (value: RelativeDatePickerValue) => void;
-  onSubmit: () => void;
-}
-
-function IntervalPicker({
-  value,
-  isNew,
-  onChange,
-  onSubmit,
-}: IntervalPickerProps) {
-  const interval = getInterval(value);
-  const options = getUnitOptions(interval);
-  const includeCurrent = getIncludeCurrent(value);
-
-  const handleIntervalChange = (inputValue: number | "") => {
-    if (inputValue !== "") {
-      onChange(setInterval(value, inputValue));
-    }
-  };
-
-  const handleUnitChange = (inputValue: string | null) => {
-    const option = options.find(option => option.value === inputValue);
-    if (option) {
-      onChange(setUnit(value, option.value));
-    }
-  };
-
-  const handleIncludeCurrentChange = () => {
-    onChange(setIncludeCurrent(value, !includeCurrent));
-  };
-
-  return (
-    <div>
-      <Flex p="md">
-        <NumberInput
-          value={interval}
-          w="4rem"
-          onChange={handleIntervalChange}
-        />
-        <Select
-          data={options}
-          value={value.unit}
-          withinPortal={false}
-          ml="md"
-          onChange={handleUnitChange}
-        />
-        <Menu withinPortal={false}>
-          <Menu.Target>
-            <Button
-              c="text.2"
-              variant="subtle"
-              leftIcon={<Icon name="ellipsis" />}
-            />
-          </Menu.Target>
-          <Menu.Dropdown>
-            <Menu.Item
-              icon={<Icon name={includeCurrent ? "check" : "calendar"} />}
-              onClick={handleIncludeCurrentChange}
-            >
-              {t`Include ${describeInterval(value.unit)}`}
-            </Menu.Item>
-          </Menu.Dropdown>
-        </Menu>
-      </Flex>
-      <Divider />
-      <Group p="sm" position="right">
-        <Button variant="filled" onClick={onSubmit}>
-          {isNew ? t`Add filter` : t`Update filter`}
-        </Button>
-      </Group>
-    </div>
   );
 }
